main:

+++ tempo/submit.go
@@ -3,37 +3,42 @@ package tempo
 import (
 	"context"
 
+	"github.com/google/go-cmp/cmp"
 	"github.com/rs/zerolog/log"
 	"github.com/tim-hilt/tempo/noteparser"
 	"github.com/tim-hilt/tempo/util"
+	"github.com/tim-hilt/tempo/util/config"
 	"golang.org/x/sync/errgroup"
 )
 
-func (t Tempo) SubmitDay(day string) {
+func (t *Tempo) SubmitDay(day string) {
 	if err := t.submit(day); err != nil {
 		log.Fatal().Err(err).Str("day", day).Msg("error when submitting")
 	}
 }
 
-func (t Tempo) submit(day string) error {
+func (t *Tempo) submit(day string) error {
+	ticketEntries, err := noteparser.ParseDailyNote(day)
+
+	if err != nil {
+		return err
+	}
+
+	// TODO: This concept doesn't quite work, when dealing with multiple files.
+	// I'd rather need a map [datestring]: []parser.DailyNoteEntry
+	// Also: It isn't relevant for submitting a single file. Maybe I can generalize this?
+	if cmp.Equal(ticketEntries, t.PreviousTicketEntries) {
+		log.Info().Msg("ticketEntries equal. not submitting")
+		return nil
+	}
+
 	worklogs, err := t.Api.FindWorklogsInRange(day, day)
 
 	if err != nil {
 		return err
 	}
 
-	// TODO: One way of submitting only changed ticket-entries would be to keep a copy of the last state of the ticketEntries and then diff their contents. Afterwards:
-	// - Deleted entries are deleted
-	// - Changed entries are deleted and re-submitted
-	// - New entries are submitted
-	// I still need to clarify how I can detect a changed ticket though
 	if err := t.Api.DeleteWorklogs(worklogs); err != nil {
-		return err
-	}
-
-	ticketEntries, err := noteparser.ParseDailyNote(day)
-
-	if err != nil {
 		return err
 	}
 
@@ -57,5 +62,12 @@ func (t Tempo) submit(day string) error {
 
 	hours, minutes := util.Divmod(workedSeconds/util.SECONDS_IN_MINUTE, util.MINUTES_IN_HOUR)
 	log.Info().Int("hours", hours).Int("minutes", minutes).Msg("successfully logged")
+
+	overtimeMinutes := (hours*util.MINUTES_IN_HOUR + minutes) - (config.GetWorkhours() * util.MINUTES_IN_HOUR)
+	overtimeHours, overtimeMinutes := util.Divmod(overtimeMinutes, util.MINUTES_IN_HOUR)
+	log.Trace().Int("overtimeHours", overtimeHours).Int("overtimeMinutes", overtimeMinutes).Msg("")
+
+	t.PreviousTicketEntries = ticketEntries
+
 	return nil
 }
